Added docs, probably did requested changes, added negative speed check for motors event.
