+++ src/main/java/aztech/modern_industrialization/compat/kubejs/MIKubeJSPlugin.java
@@ -26,7 +26,6 @@ package aztech.modern_industrialization.compat.kubejs;
 import aztech.modern_industrialization.MIStartup;
 import aztech.modern_industrialization.compat.kubejs.machine.MIMachineKubeJSEvents;
 import aztech.modern_industrialization.compat.kubejs.material.MIMaterialKubeJSEvents;
-import aztech.modern_industrialization.compat.kubejs.motor.MIPipeKubeJSEvents;
 import aztech.modern_industrialization.compat.kubejs.recipe.MIRecipeKubeJSEvents;
 import aztech.modern_industrialization.compat.kubejs.recipe.MachineRecipeSchema;
 import aztech.modern_industrialization.compat.kubejs.registration.MIRegistrationKubeJSEvents;
@@ -42,7 +41,6 @@ public class MIKubeJSPlugin extends KubeJSPlugin {
 
     @Override
     public void registerEvents() {
-        MIPipeKubeJSEvents.EVENT_GROUP.register();
         MIMachineKubeJSEvents.EVENT_GROUP.register();
         MIMaterialKubeJSEvents.EVENT_GROUP.register();
         MIRecipeKubeJSEvents.EVENT_GROUP.register();
